Added PetPage.html (& Photos)
Added the first draft of the PetPage.
PetPage is currently an html shell, still needs to be updated with php and database info.

Also added some additional photos of cats that Katie knows. Definetly all of them are cats and none of them are a raccoon that once showed up on her porch.